Fixing message in saved-search alert command and execution on windows by removing shebang statement
Splunk 6.0.1 tries to parse the shebang and execute the script using the command specified, however /usr/bin/env doesn't exist on Windows so this fails. If the shebang is removed Splunk falls back to determining the environment to run based on the file extension. This fix doesn't work on Splunk 6.0.0 or older so I'm not keen on it but I can't see any other portable ways to fix this.